Change search to tokenize input and search w/ OR
Seeing lots of search traffic for things like "montauk, nj", or
"seldovia ak", for which there are actually entries.  Prior algorithm
wouldn't show any search results because it was an exact substring match
search.

Now we tokenize on [ ,]+ and just OR the results -- also not what the
user exactly wants, but at least it shows there are results, and it
should lead them to refine/narrow their search by removing the
extraneous terms.