Convert caching to use sliding +/- 12 month window
Current implementation suffers from an EOY data starvation problem: by
caching year to year, we have no data to show on 12/31, prompting folks
to think the system might be broken / not functioning properly
(practically speaking, they are right).

Now we slide a window around the given date (defaulting to the current
month), and cache the data for that time range, per month.  This means
we calculate (and cache) the calendar 12x a year (vs. previously 1x),
but this is More Correct, compute is cheap and disk is even cheaper.