Change sliding time window behavior to minimize the past
Used to be a simple +/- N months; this changes it to 1 month prior +
whatever is left of the window size (- last month - this month) in the
future.

For NOAA currents + CHS tides, the limit is 366 days.  So this is 10
months in the future.  For NOAA tides, the limit is at least under 2
years, so we do 14 months -- 1 behind, this month, and 12 in the future.