Remove bogus CHS stations once we detect they are bogus
So this happened: station 5cebf1e23d0f4a073c4bbfb4
returned empty data.  It is type: DISCONTINUED,
operating: false, which would strongly imply we
should filter those out from the list, however a
different station 5cebf1df3d0f4a073c4bbcb9 also
has the same type/operating but *DOES* return tide
data.

Upon further research, it turns out the CHS
metadata is a fucking mess.  Of the tons of
possible indicators in the metadata, none are
reliable as way to know if an "active" server in
the station list will return actual tide data.
This is to say, some stations that are Temporary,
Discontinued, or operating:false, etc, will return
data, while others that are status:OK, Permanent,
etc, won't.  The only way to know is to try to
retrieve it.

Which means we can't filter them out of the list
ahead of time, since there are API ratelimits and
over 1k stations to double-check.  All we can do
is (1) return nil to the caller, and (2) nuke the
station from the list post-facto.  This means bad
stations will show up in the search results until
someone attempts to use one -- then it will get
nuked.

Super lame, but what else can we practically do?